+++ CHANGES.md
@@ -14,6 +14,7 @@
 * Built using Avro 1.9.1
 * GenerateAvroProtocolTask now has a `classpath` property; defaults to the runtime configuration when the Avro plugin is applied
 * GenerateAvroProtocolTask now properly declares the `classpath` as an input; fixes #86; thanks to [RichSteele](https://github.com/RichSteele) for the bug report
+* Fix handling of default `outputCharacterEncoding` (use of system default character set to match Java compiler)
 
 ## 0.17.0
 * Built using Avro 1.9.0